Move instrumentation helpers to Appsignal.Instrumentation
